improve some things
