+++ Zavrazhin/lab2/Makefile
@@ -1,2 +1,2 @@
 lab1: Source/main.cpp Source/expression.h
-	g++ Source/main.cpp -std=c++17 -o lab2
+	g++ Source/main.cpp -std=c++17 -fconcepts -o lab2
